split lib/ into submodules
to make code base easier to navigate and understand
