+++ src/lib/element/index.ts
@@ -1,0 +1,4 @@
+export { default as ElementHeading } from './ElementHeading.svelte'
+export { default as ElementPhoto } from './ElementPhoto.svelte'
+export { default as ElementStats } from './ElementStats.svelte'
+export { default as ElementTile } from './ElementTile.svelte'
