+++ src/lib/index.ts
@@ -2,26 +2,15 @@ import type { categories, elem_symbols } from './labels'
 
 export { default as BohrAtom } from './BohrAtom.svelte'
 export { default as ColorBar } from './ColorBar.svelte'
-export { default as ColorCustomizer } from './ColorCustomizer.svelte'
 export { default as ColorScaleSelect } from './ColorScaleSelect.svelte'
-export { default as ElementHeading } from './ElementHeading.svelte'
-export { default as ElementPhoto } from './ElementPhoto.svelte'
-export { default as ElementScatter } from './ElementScatter.svelte'
-export { default as ElementStats } from './ElementStats.svelte'
-export { default as ElementTile } from './ElementTile.svelte'
 export { default as Icon } from './Icon.svelte'
-export { default as Line } from './Line.svelte'
 export { default as Nucleus } from './Nucleus.svelte'
-export { default as PeriodicTable, default } from './PeriodicTable.svelte'
-export { default as PropertySelect } from './PropertySelect.svelte'
-export { default as ScatterPlot } from './ScatterPlot.svelte'
-export { default as ScatterPoint } from './ScatterPoint.svelte'
-export { default as Structure } from './Structure.svelte'
-export { default as StructureCard } from './StructureCard.svelte'
-export { default as StructureLegend } from './StructureLegend.svelte'
-export { default as StructureScene } from './StructureScene.svelte'
-export { default as TableInset } from './TableInset.svelte'
-export { default as element_data } from './element-data'
+export * from './element'
+export { default as element_data } from './element/data'
+export * from './labels'
+export * from './ptable'
+export * from './scatter'
+export { default as ScatterPoint } from './scatter/ScatterPoint.svelte'
 export * from './structure'
 
 export type Category = (typeof categories)[number]
