+++ src/lib/ptable/index.ts
@@ -1,0 +1,4 @@
+export { default as ColorCustomizer } from './ColorCustomizer.svelte'
+export { default as PeriodicTable, default } from './PeriodicTable.svelte'
+export { default as PropertySelect } from './PropertySelect.svelte'
+export { default as TableInset } from './TableInset.svelte'
