+++ src/lib/scatter/ScatterPoint.svelte
