+++ src/lib/scatter/index.ts
@@ -1,0 +1,3 @@
+export { default as ElementScatter } from './ElementScatter.svelte'
+export { default as Line } from './Line.svelte'
+export { default as ScatterPlot } from './ScatterPlot.svelte'
