+++ src/lib/stores.ts
@@ -1,6 +1,6 @@
+import type { Category, ChemicalElement } from '$lib'
 import { session_store } from 'svelte-zoo/stores'
 import { writable } from 'svelte/store'
-import type { Category, ChemicalElement } from '.'
 import { default_category_colors, default_element_colors } from './colors'
 
 export const active_category = writable<Category | null>(null)
