+++ tests/periodic-table.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test'
-import element_data from '../src/lib/element-data.ts'
+import element_data from '../src/lib/element/data.js'
 import {
   categories,
   category_counts,
